requirements.txt
beautifulsoup4==4.12.3
cbor==1.0.0
certifi==2024.12.14
charset-normalizer==3.4.0
chest==0.2.3
colaburl==0.1.0
Deprecated==1.2.15
dill==0.3.9
filelock==3.16.1
fire==0.7.0
fsspec==2024.12.0
HeapDict==1.0.1
huggingface-hub==0.27.0
idna==3.10
ijson==3.3.0
inscriptis==2.5.0
ir_datasets==0.5.9
ir_measures==0.3.6
Jinja2==3.1.5
joblib==1.4.2
lxml==5.3.0
lz4==4.3.3
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.4.2
npids==0.1.1
numpy==1.26.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.0.0
protobuf==5.29.3
pyciff==0.1.1
pyjnius==1.6.1
pyterrier-adaptive==0.2.0
pyterrier-alpha==0.12.1
pyterrier-dr==0.2.0
pyterrier_pisa==0.2.1
python-dateutil==2.9.0.post0
python-terrier==0.13.0
pytrec-eval-terrier==0.5.6
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
sentence-transformers==3.3.1
six==1.17.0
soupsieve==2.6
statsmodels==0.14.4
sympy==1.13.1
tbb==2020.3.254
termcolor==2.5.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.47.1
trec-car-tools==2.6
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
unlzw3==0.2.3
urllib3==2.3.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
wget==3.2
wrapt==1.17.0
zlib-state==0.1.9
